+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDynamicPanningData1.js
@@ -8,22 +8,23 @@ const bubbleDynamicPanningData = {
   },
   onClick: loadBubblePopup,
   color: Carbon.helpers.COLORS.PINK,
-  values: [{
-    x: new Date(2016, 0, 1, 12, 0).toISOString(),
-    y: 20,
-  },
-  {
-    x: new Date(2016, 0, 1, 15, 45).toISOString(),
-    y: 40,
-  },
-  {
-    x: new Date(2016, 0, 1, 10, 30).toISOString(),
-    y: 50,
-  },
-  {
-    x: new Date(2016, 0, 1, 18, 0).toISOString(),
-    y: 60,
-  },
+  values: [
+    {
+      x: new Date(2016, 0, 1, 12, 0).toISOString(),
+      y: 70,
+    },
+    {
+      x: new Date(2016, 0, 1, 15, 45).toISOString(),
+      y: 120,
+    },
+    {
+      x: new Date(2016, 0, 1, 10, 30).toISOString(),
+      y: 160,
+    },
+    {
+      x: new Date(2016, 0, 1, 18, 0).toISOString(),
+      y: 180,
+    },
   ],
   yAxis: 'y',
   showShapes: false,