+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/BubbleGraph/Reflow/MultipleDatasets.jsx
@@ -4,33 +4,32 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import '@cerner/terra-graphs/lib/components/Graph.module.scss';
 import Button from 'terra-button/lib/Button';
-import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/simplePanningAxisData';
-import bubbleDynamicPanningData from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDynamicPanningData.js';
-import bubbleDynamicPanningData2 from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDynamicPanningData2.js';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bubble/simplePanningAxisData';
+import bubbleDataBasic from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bubble/bubbleDataBasic.js';
+import bubbleDataCustomRadius from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bubble/bubbleDataCustomRadius.js';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const dataset1 = utils.deepClone(bubbleDynamicPanningData[0]);
-const dataset2 = utils.deepClone(bubbleDynamicPanningData2[0]);
+const graphConfig = utils.deepClone(getSimpleAxisData('#BubbleMultipleDatasetsPanning'));
+const dataset1 = utils.deepClone(bubbleDataBasic);
+const dataset2 = utils.deepClone(bubbleDataCustomRadius);
 
-const initialState = {
+const state = {
   initial: 0,
   factor: 3,
-  graphConfig: utils.deepClone(getSimpleAxisData('#BubbleMultipleDatasetsPanning')),
 };
 
 let graph;
 
 const BubblePanningExample = () => {
   React.useEffect(() => {
-    graph = Carbon.api.graph(initialState.graphConfig);
+    graph = Carbon.api.graph(graphConfig);
     graph.loadContent(Carbon.api.bubbleMultipleDataset(dataset1));
     graph.loadContent(Carbon.api.bubbleMultipleDataset(dataset2));
   }, []);
 
   const reducer = (panState, action) => {
-    const newDatasets = { panData: [utils.deepClone(bubbleDynamicPanningData[1]), utils.deepClone(bubbleDynamicPanningData2[1])] };
     let hour;
 
     switch (action.type) {
@@ -44,18 +43,22 @@ const BubblePanningExample = () => {
         return panState;
     }
 
-    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
-
-    graph.reflowMultipleDatasets(newDatasets);
-
     return {
       initial: hour,
       factor: panState.factor,
     };
   };
 
-  const [, dispatch] = React.useReducer(reducer, initialState);
+  const [panState, dispatch] = React.useReducer(reducer, state);
+
+  React.useLayoutEffect(() => {
+    if (!graph) { return; }
+
+    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, panState.initial).toISOString();
+    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, panState.initial).toISOString();
+
+    graph.reflowMultipleDatasets();
+  }, [panState.initial]);
 
   return (
     <React.Fragment>