+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/BubbleGraph/Reflow/MultipleDatasets.jsx
@@ -3,15 +3,14 @@ import BubbleMultipleDataset from '@cerner/terra-graphs/lib/components/Bubble/Bu
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/simplePanningAxisData';
-import bubbleDataBasic from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataBasic.js';
-import bubbleDataCustomRadius from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataCustomRadius.js';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bubble/simplePanningAxisData';
+import bubbleDataBasic from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bubble/bubbleDataBasic.js';
+import bubbleDataCustomRadius from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bubble/bubbleDataCustomRadius.js';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const basicDataset = utils.deepClone(bubbleDataBasic);
-const customRadiusDataset = utils.deepClone(bubbleDataCustomRadius);
+const dataset = [utils.deepClone(bubbleDataBasic), utils.deepClone(bubbleDataCustomRadius)];
 
 const initialState = {
   initial: 0,
@@ -46,13 +45,13 @@ const BubblePanningExample = () => {
   };
 
   const [panState, dispatch] = React.useReducer(reducer, initialState);
-  const bubbleData = [basicDataset, customRadiusDataset];
+
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <BubbleMultipleDataset graphID="BubbleMultipleDatasetsPanning" graphConfig={panState.graphConfig} dataset={bubbleData} />
+      <BubbleMultipleDataset graphID="BubbleMultipleDatasetsPanning" graphConfig={panState.graphConfig} dataset={dataset} />
     </React.Fragment>
   );
 };